Update Tables.html
I posted a comment about this earlier on the website. The asymptotic notation for a bunch of these are technically insufficient and the base for a bunch of these log's are also incorrect. Lg should be used for log base 2 and theta should be used when there is indeed a tight asymptotic bound. This book (https://www.cs.duke.edu/courses/fall08/cps230/Book.pdf) is a good source to see what i'm talking about if you're unsure when to use big O, theta, or omega.